requirements.txt
Babel==2.12.1
blinker==1.7.0
cachelib==0.10.2
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-Session==0.5.0
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.7.0
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
zipp==3.17.0
